prep for 1.39 (#3128)

+++ modules/mediawiki/files/bin/mwupgradetool.py
@@ -2,9 +2,9 @@
 
 import os
 import requests
+import sys
 canary = 'mwtask141'
-hatedserver = 'mw131'
-serverlist = 'mw121,mw122,mw132,mw141,mw142'
+serverlist = 'mw121,mw122,mw131,mw132,mw141,mw142'
 
 
 def check_up(server: str) -> bool:
@@ -15,51 +15,25 @@ def check_up(server: str) -> bool:
     return False
 
 
-def check_ro(server: str) -> bool:
-    headers = {'X-Miraheze-Debug': f'{server}.miraheze.org'}
-    req = requests.get('https://meta.miraheze.org/w/api.php?action=query&meta=siteinfo&formatversion=2&format=json', headers=headers)
-    response = req.json()
-    if response['query']['general']['readonly']:
-        return True
-    return False
-
-
 print('Welcome to the MediaWiki Upgrade tool!')
 input('Please confirm you are running this script on the canary server: (press enter)')
-print('MediaWiki -> RO - Syncing config')
-os.system('deploy-mediawiki --config --pull=config --servers=all')
-print('Config deployed')
-print('Checking RO on Canary Server')
-if not check_ro(canary):
-    input('Stopping deploy - RO check failed - Press enter to resume')
-for server in ['mw121', 'mw122', 'mw131', 'mw132', 'mw141', 'mw142']:
-    print(f'Confirming RO on {server}')
-    if not check_ro(server):
-        input(f'RO check failed on {server} - Press enter to resume')
-print('Starting staging update')
-input('Press enter when branch updated in puppet: ')
-os.system('sudo -u www-data rm -rf /srv/mediawiki-staging/w')
-os.system('sudo puppet agent -tv')
-print('Will now check mediawiki branch')
-os.system('git -C /srv/mediawiki-staging/w rev-parse --abbrev-ref HEAD')
-input('Confirm: ')
-print('Will now deploy to canary server')
-os.system(f'deploy-mediawiki --world --l10n --force --ignore-time --extension-list --servers={canary}')
-if check_up(canary) and check_ro(canary):
-    print('Canary deploy done')
-else:
-    print('Canary is not online')
-input('Press enter to rollout: ')
-print(f'Deploying to {hatedserver}')
-os.system(f'deploy-mediawiki --world --l10n --force --ignore-time --extension-list --servers={hatedserver}')
-if check_up(hatedserver) and check_ro(hatedserver):
-    print(f'{hatedserver} deploy done')
-else:
-    print(f'{hatedserver} is not online')
-input('Confirm mass rolout:')
-os.system(f'deploy-mediawiki --world --l10n --force --ignore-time --extension-list --servers={serverlist}')
-print('Deployment done')
-input('Please merge RW change and press enter: ')
-print('Running puppet')
-os.system('sudo puppet agent -tv')
-print('Deployment done')
+if sys.argv[1] == 'prep':
+    print('Starting staging update')
+    input('Press enter when branch updated in puppet: ')
+    os.system('sudo -u www-data rm -rf /srv/mediawiki-staging/w')
+    os.system('sudo puppet agent -tv')
+    print('Will now check mediawiki branch')
+    os.system('git -C /srv/mediawiki-staging/w rev-parse --abbrev-ref HEAD')
+    input('Confirm: ')
+    print('Will now deploy to canary server')
+    os.system(f'deploy-mediawiki --world --l10n --force --ignore-time --extension-list --servers={canary}')
+    if check_up(canary):
+        print('Canary deploy done')
+    else:
+        print('Canary is not online')
+    print('Deployment done, run with "fleet" to rollout')
+if sys.argv[1] == 'fleet':
+    input('Press enter when all servers pooled: ')
+    input('Confirm mass rolout:')
+    os.system(f'deploy-mediawiki --world --l10n --force --ignore-time --extension-list --servers={serverlist}')
+    print('Deployment done')
